add tests for player-factory function and fix linter errors
