--- __mocks__/document.js
@@ -1,11 +1,0 @@
-// const document = () => {
-const querySelector = (selector) => {
-  if (selector === '.system-message') {
-    return {
-      innerHTML: '',
-    };
-  }
-};
-// }
-
-exports.querySelector = querySelector;
