+++ main.js
@@ -2,27 +2,7 @@ import { Players } from './players';
 import { gameBoard } from './gameboard';
 import { logic } from './logic';
 import { UI } from './UI';
-
-const playerFactory = (playerName, playerSymbol) => {
-  const name = playerName;
-  const symbol = playerSymbol;
-  let playerMoves = [];
-  const getName = () => name;
-  const getSymbol = () => symbol;
-  const getPlayerMoves = () => playerMoves;
-  const addMove = (move) => playerMoves.push(move);
-  const resetPlayerMoves = () => {
-    playerMoves = [];
-  };
-
-  return {
-    getName,
-    getSymbol,
-    getPlayerMoves,
-    addMove,
-    resetPlayerMoves,
-  };
-};
+import { playerFactory } from './player-factory';
 
 const game = (() => {
   const player1 = playerFactory('One', 'X');
